pyproject.toml
[tool.poetry]
name = "socialnetwork"
version = "0.1.0"
description = "A social network experiment"
authors = ["Manuel Pizano <manuel_pizano@maniakapps.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.7"
Faker-Events = "^1.6.0"
Pillow = "^9.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
